+++ packages/ripple/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/ripple
 
+## 1.1.0
+
+### Minor Changes
+
+- 65032024: Updates what options are able to be passed to Ripple component
+
+### Patch Changes
+
+- Updated dependencies [65032024]
+  - @leafygreen-ui/palette@3.2.0
+
 ## 1.0.0
 
 ### Major Changes